cover multi-file generic enum method layout
