+++ tests/integration/cli_build/frontend_json/layout_golden/generic.rs
@@ -38,6 +38,15 @@ fn emit_json_layout_generic_enum_method_nested_result_schema_matches_golden() {
 }
 
 #[test]
+fn emit_json_layout_multi_file_generic_enum_method_schema_matches_golden() {
+    assert_layout_matches_fixture(
+        &fixture("tests/zen/multi_file_generic_enum_method/main.zen"),
+        "multi-file generic enum method program input",
+        "tests/fixtures/ir_json/layout_multi_file_generic_enum_method.golden.json",
+    );
+}
+
+#[test]
 fn emit_json_layout_multi_file_generic_method_nested_result_schema_matches_golden() {
     assert_layout_matches_fixture(
         &fixture("tests/zen/multi_file_type_method_nested_result_dependency/main.zen"),
